// platform/app/src/telepacs/api/generated/model/accountStatusEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TelePACS Backend API
 * Versioned API contract for the independent TelePACS backend.
 * OpenAPI spec version: 0.1.0
 */

/**
 * * `PENDING_VERIFICATION` - Pending verification
 * * `ONBOARDING` - Onboarding
 * * `PENDING_REVIEW` - Pending review
 * * `ACTIVE` - Active
 * * `REJECTED` - Rejected
 * * `SUSPENDED` - Suspended
 * * `CLOSED` - Closed
 */
export type AccountStatusEnum = typeof AccountStatusEnum[keyof typeof AccountStatusEnum];


export const AccountStatusEnum = {
  PENDING_VERIFICATION: 'PENDING_VERIFICATION',
  ONBOARDING: 'ONBOARDING',
  PENDING_REVIEW: 'PENDING_REVIEW',
  ACTIVE: 'ACTIVE',
  REJECTED: 'REJECTED',
  SUSPENDED: 'SUSPENDED',
  CLOSED: 'CLOSED',
} as const;